requirements.txt
fastapi
uvicorn[standard]

sqlalchemy[asyncio]
asyncpg
alembic
psycopg[binary,pool]

pydantic
pydantic-settings
python-dotenv

passlib[bcrypt]
bcrypt
python-jose[cryptography]

celery[redis]
redis

httpx
python-dateutil

openai

google-api-python-client
google-auth
google-auth-oauthlib

email-validator
python-multipart

pytest
pytest-asyncio
